+++ tests/pageObjects/sort-bar.ts
@@ -1,19 +1,6 @@
 import { type Page, type Locator, expect } from '@playwright/test';
 
-export type SortFilter = 
-  | 'Weekly views' 
-  | 'All-time views' 
-  | 'Title' 
-  | 'Date published'
-  | 'Date archived'
-  | 'Date reviewed'
-  | 'Date added'
-  | 'Creator';
-export type SortOrder = 'ascending' | 'descending';
-export type DateMetadataLabel = {
-  filter: string,
-  date: string
-}
+import { SortOrder } from '../models';
 
 export class SortBar {
   readonly page: Page;
@@ -30,22 +17,6 @@ export class SortBar {
     this.sortSelector = this.sortFilterBar.locator('ul#desktop-sort-selector');
     this.btnSortDirection = this.sortFilterBar.locator('.sort-direction-icon');
     this.srSortText = this.sortFilterBar.locator('button.sort-direction-selector span.sr-only');
-  }
-
-  viewsSorted(order: SortOrder, arr: Number[]) {
-    if (order === 'ascending') {
-      return arr.every((x, i) => i === 0 || x >= arr[i - 1]);
-    } else {
-      return arr.every((x, i) => i === 0 || x <= arr[i - 1]);
-    }
-  }
-
-  datesSorted(order: SortOrder, arr: DateMetadataLabel[]) {
-    if (order === 'ascending') {
-      return arr.every((x, i) => i === 0 || new Date(x.date) >= new Date(arr[i - 1].date));
-    } else {
-      return arr.every((x, i) => i === 0 || new Date(x.date) <= new Date(arr[i - 1].date));
-    }
   }
 
   async buttonClick (sortName: string) {
